+++ UnityEditor.TestRunner/TestRun/Tasks/BuildTestTreeTask.cs
@@ -15,7 +15,6 @@ namespace UnityEditor.TestTools.TestRunner.TestRun.Tasks
         public BuildTestTreeTask(TestPlatform testPlatform)
         {
             m_TestPlatform = testPlatform;
-            RerunAfterResume = true;
         }
 
         internal IEditorLoadedTestAssemblyProvider m_testAssemblyProvider = new EditorLoadedTestAssemblyProvider(new EditorCompilationInterfaceProxy(), new EditorAssembliesProxy());