+++ UnityEditor.TestRunner/TestRunner/EditModeRunner.cs
@@ -41,7 +41,7 @@ namespace UnityEditor.TestTools.TestRunner
         private int m_CurrentPC;
 
         [SerializeField]
-        public bool m_ResumeAfterDomainReload;
+        private bool m_ExecuteOnEnable;
 
         [SerializeField]
         private List<string> m_AlreadyStartedTests;
@@ -52,10 +52,17 @@ namespace UnityEditor.TestTools.TestRunner
         [SerializeField]
         private List<ScriptableObject> m_CallbackObjects = new List<ScriptableObject>();
 
-        private TestStartedEvent m_TestStartedEvent;
-        private TestFinishedEvent m_TestFinishedEvent;
-        private RunStartedEvent m_RunStartedEvent;
-        private RunFinishedEvent m_RunFinishedEvent;
+        [SerializeField]
+        private TestStartedEvent m_TestStartedEvent = new TestStartedEvent();
+
+        [SerializeField]
+        private TestFinishedEvent m_TestFinishedEvent = new TestFinishedEvent();
+
+        [SerializeField]
+        private RunStartedEvent m_RunStartedEvent = new RunStartedEvent();
+
+        [SerializeField]
+        private RunFinishedEvent m_RunFinishedEvent = new RunFinishedEvent();
 
         [SerializeField]
         private TestRunnerStateSerializer m_TestRunnerStateSerializer = new TestRunnerStateSerializer();
@@ -87,41 +94,37 @@ namespace UnityEditor.TestTools.TestRunner
 
         public IUnityTestAssemblyRunnerFactory UnityTestAssemblyRunnerFactory { get; set; }
 
-        public void Init(Filter[] filters, TestPlatform platform, bool runningSynchronously, ITest testTree, 
-            RunStartedEvent runStartedEvent, TestStartedEvent testStartedEvent, TestFinishedEvent testFinishedEvent, RunFinishedEvent runFinishedEvent)
+        public void Init(Filter[] filters, TestPlatform platform, bool runningSynchronously)
         {
             m_Filters = filters;
             m_TestPlatform = platform;
             m_AlreadyStartedTests = new List<string>();
             m_ExecutedTests = new List<TestResultSerializer>();
             RunningSynchronously = runningSynchronously;
-            m_RunStartedEvent = runStartedEvent;
-            m_TestStartedEvent = testStartedEvent;
-            m_TestFinishedEvent = testFinishedEvent;
-            m_RunFinishedEvent = runFinishedEvent;
-            InitRunner(testTree);
+            InitRunner();
         }
 
-        private void InitRunner(ITest testTree)
+        private void InitRunner()
         {
             //We give the EditMode platform here so we dont suddenly create Playmode work items in the test Runner.
             m_Runner = (UnityTestAssemblyRunnerFactory ?? new UnityTestAssemblyRunnerFactory()).Create(TestPlatform.EditMode, new EditmodeWorkItemFactory());
-            m_Runner.LoadTestTree(testTree);
+            var testAssemblyProvider = new EditorLoadedTestAssemblyProvider(new EditorCompilationInterfaceProxy(), new EditorAssembliesProxy());
+            var assemblies = testAssemblyProvider.GetAssembliesGroupedByType(m_TestPlatform).Select(x => x.Assembly).ToArray();
+            var loadedTests = m_Runner.Load(assemblies, TestPlatform.EditMode,
+                UnityTestAssemblyBuilder.GetNUnitTestBuilderSettings(m_TestPlatform));
+            loadedTests.ParseForNameDuplicates();
+            CallbacksDelegator.instance.TestTreeRebuild(loadedTests);
             hideFlags |= HideFlags.DontSave;
             EnumerableSetUpTearDownCommand.ActivePcHelper = new EditModePcHelper();
             OuterUnityTestActionCommand.ActivePcHelper = new EditModePcHelper();
         }
 
-        public void Resume(ITest testTree, RunStartedEvent runStartedEvent, TestStartedEvent testStartedEvent, TestFinishedEvent testFinishedEvent, RunFinishedEvent runFinishedEvent)
+        public void OnEnable()
         {
-            if (m_ResumeAfterDomainReload)
+            if (m_ExecuteOnEnable)
             {
-                m_RunStartedEvent = runStartedEvent;
-                m_TestStartedEvent = testStartedEvent;
-                m_TestFinishedEvent = testFinishedEvent;
-                m_RunFinishedEvent = runFinishedEvent;
-                InitRunner(testTree);
-                m_ResumeAfterDomainReload = false;
+                InitRunner();
+                m_ExecuteOnEnable = false;
                 foreach (var callback in m_CallbackObjects)
                 {
                     AddListeners(callback as ITestRunnerListener);
@@ -184,11 +187,22 @@ namespace UnityEditor.TestTools.TestRunner
             var testListenerWrapper = new TestListenerWrapper(m_TestStartedEvent, m_TestFinishedEvent);
             m_RunStep = m_Runner.Run(testListenerWrapper, GetFilter()).GetEnumerator();
             m_RunningTests = true;
+
+            if (!RunningSynchronously) 
+                EditorApplication.update += TestConsumer;
+        }
+
+        public void CompleteSynchronously()
+        {
+            while (!m_Runner.IsTestComplete)
+                TestConsumer();
         }
 
         private void OnBeforeAssemblyReload()
         {
-            if (m_ResumeAfterDomainReload)
+            EditorApplication.update -= TestConsumer;
+
+            if (m_ExecuteOnEnable)
             {
                 AssemblyReloadEvents.beforeAssemblyReload -= OnBeforeAssemblyReload;
                 return;
@@ -248,7 +262,7 @@ namespace UnityEditor.TestTools.TestRunner
             return MoveNextAndUpdateYieldObject();
         }
 
-        public void TestConsumer()
+        private void TestConsumer()
         {
             var moveNext = MoveNextAndUpdateYieldObject();
 
@@ -271,6 +285,9 @@ namespace UnityEditor.TestTools.TestRunner
 
         private void CompleteTestRun()
         {
+            if (!RunningSynchronously)
+                EditorApplication.update -= TestConsumer;
+   
             TestLauncherBase.ExecutePostBuildCleanupMethods(this.GetLoadedTests(), this.GetFilter(), Application.platform);
             
             m_RunFinishedEvent.Invoke(m_Runner.Result);
@@ -357,7 +374,7 @@ namespace UnityEditor.TestTools.TestRunner
         {
             m_TestRunnerStateSerializer.SaveContext();
             m_CurrentPC = EnumeratorStepHelper.GetEnumeratorPC(TestEnumerator.Enumerator);
-            m_ResumeAfterDomainReload = true;
+            m_ExecuteOnEnable = true;
 
             RunningTests = false;
         }
@@ -384,6 +401,7 @@ namespace UnityEditor.TestTools.TestRunner
         public void Dispose()
         {
             Reflect.MethodCallWrapper = null;
+            EditorApplication.update -= TestConsumer;
 
             DestroyImmediate(this);
 
@@ -396,12 +414,13 @@ namespace UnityEditor.TestTools.TestRunner
                 m_CallbackObjects.Clear();
             }
             RunningTests = false;
+            EditorApplication.UnlockReloadAssemblies();
         }
 
         public void OnRunCancel()
         {
             UnityWorkItemDataHolder.alreadyExecutedTests = null;
-            m_ResumeAfterDomainReload = false;
+            m_ExecuteOnEnable = false;
             m_Runner.StopRun();
             RunFinished = true;
         }