+++ UnityEditor.TestRunner/UnityTestProtocol/UtpDebuglogger.cs
@@ -4,12 +4,10 @@ namespace UnityEditor.TestTools.TestRunner.UnityTestProtocol
 {
     class UtpDebugLogger : IUtpLogger
     {
-        internal const string UtpPrefix = "\n##utp:";
-
         public void Log(Message msg)
         {
             var msgJson = JsonUtility.ToJson(msg);
-            Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null, "{0}{1}", UtpPrefix, msgJson);
+            Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null, "\n##utp:{0}", msgJson);
         }
     }
 }